backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Rizki Perdana",email = "perdana.rizki16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "fastapi-cache (>=0.1.0,<0.2.0)",
    "aerich (>=0.9.1,<0.10.0)",
    "websocket (>=0.2.1,<0.3.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "asgi-correlation-id (>=4.3.4,<5.0.0)",
    "fastapi-pagination (>=0.13.3,<0.14.0)",
    "orjson (>=3.11.2,<4.0.0)",
    "ruff (>=0.12.9,<0.13.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "tomlkit (>=0.13.3,<0.14.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.core.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
